add prepare_database task
This task prepares the MySQL database for deploying. This means the
client MySQL configuration is now a template and the values inside are
in the Rexfile. We might need to change this to a configuration file
soon. We might also want to make the defaults file into something that
is always loaded forever.