preserve host header through proxy to fix urls
It's impossible to build a URL with the correct host if we don't know
the host the user is actually requesting. By preserving the `Host:` HTTP
header, we get the right information. This fixes especially the API
documentation site (http://api.cpantesters.org/docs) to make requests to
the right place.