fix distributions with `+` in their name
Text-Tabs+Wrap is a valid distribution (PAUSE controls what that means,
not CPAN Testers), and it is now correctly passing through our
RewriteRules. It is not correctly handled by Labyrinth, so the
www.cpantesters.org webapp does not show it correctly. It _IS_ correctly
handled by the static JSON files served by Apache, so
matrix.cpantesters.org can show this distribution correctly. That was
acceptable to Aristotle, so it's acceptable to me.