update backpan to update thrice a day
The backpan.sh script takes about 4 hours to run. This is probably due
to me not configuring affinities for the persistent volumes used by CPAN
and BackPAN and the various jobs to ensure that they all run close
together in the same DC if not the same machine. But Neil Bowers said
that once a day would be plenty, so I'm gonna compromise and do it about
every 8 hours.